# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U1R API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product list",
                "operationId": "8a8390dc6232a161e105158a556efa18",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id"
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Brand Id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Data Successfull"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a product",
                "operationId": "59e7130032df1fa48d49ca988351efae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "cros_out_price",
                                    "main_category_id",
                                    "sub_category_id",
                                    "category_id",
                                    "brand_id",
                                    "store_id"
                                ],
                                "properties": {
                                    "name": {},
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "cros_out_price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "description": {},
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "image": {}
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "main_category_id": {
                                        "type": "integer"
                                    },
                                    "sub_category_id": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "store_id": {
                                        "type": "integer"
                                    },
                                    "variants": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "color_id": {
                                                    "type": "integer"
                                                },
                                                "size_id": {
                                                    "type": "integer"
                                                },
                                                "stock": {
                                                    "type": "integer"
                                                },
                                                "image": {}
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product with id",
                "operationId": "63cb35ad009c1a01ef3e6be17473236c",
                "responses": {
                    "200": {
                        "description": "Get Data Successfull"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "ef1e562bd1840fd6fadf8b1439f6cee5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "cros_out_price",
                                    "main_category_id",
                                    "sub_category_id",
                                    "category_id",
                                    "brand_id",
                                    "store_id"
                                ],
                                "properties": {
                                    "name": {},
                                    "price": {
                                        "type": "integer"
                                    },
                                    "cros_out_price": {
                                        "type": "integer"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "description": {},
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "image": {}
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "main_category_id": {
                                        "type": "integer"
                                    },
                                    "sub_category_id": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "store_id": {
                                        "type": "integer"
                                    },
                                    "variants": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "color_id": {
                                                    "type": "integer"
                                                },
                                                "size_id": {
                                                    "type": "integer"
                                                },
                                                "stock": {
                                                    "type": "integer"
                                                },
                                                "image": {}
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}